Rename app and register basic admin
