+++ web-app/django/VIM/apps/instruments/models/__init__.py
@@ -1,4 +1,4 @@
 from VIM.apps.instruments.models.instrument import Instrument
 from VIM.apps.instruments.models.instrument_name import InstrumentName
 from VIM.apps.instruments.models.language import Language
-from VIM.apps.instruments.models.avresource import AVResource
+from VIM.apps.instruments.models.av_resource import AVResource
