+++ web-app/django/VIM/apps/vim/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class VimConfig(AppConfig):
+class InstrumentsConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "vim"
+    name = "instruments"
