--- web-app/django/VIM/apps/vim/models/__init__.py
@@ -1,4 +1,0 @@
-from VIM.apps.vim.models.instrument import Instrument
-from VIM.apps.vim.models.instrument_name import InstrumentName
-from VIM.apps.vim.models.language import Language
-from VIM.apps.vim.models.av_resource import AVResource
