--- web-app/django/VIM/apps/vim/models/language.py
@@ -1,12 +1,0 @@
-from django.db import models
-
-
-class Language(models.Model):
-    wikidata_code = models.CharField(max_length=10, unique=True, blank=False)
-    wikidata_id = models.IntegerField(unique=True, primary_key=True)
-    en_label = models.CharField(
-        max_length=50, blank=False, help_text="Language label in English"
-    )
-    autonym = models.CharField(
-        max_length=50, blank=False, help_text="Language label in the language itself"
-    )
